\name{Series2Data}

\alias{Series2Data}


\alias{CobbDouglas}
\alias{logCobbDouglas}

\alias{Greene4Table131}

\alias{pr45}

\alias{spc1970}
\alias{spcindis}


\title{fSeries Data Sets} 


\description{

    A collection and description of data sets 
    used in examples of multivariate data analysis.
    \cr

    \code{CobbDouglas}
    \code{logCobbDouglas}
    \cr
    This are the data for the Cobb-Douglas productivity function
    as used in exaple 7.10 in the book of D.N. Gujarati. The three
    colummns of the annual data set ranging from 1958 until 1972
    are, Y the output, X2 the labor input, and X3 the capital input.
    
    \code{Greene4Table131}
    \cr
    US yearly investment data from 1963 until 1982 as listed in Table
    A13.1 in the book of W. Greene, Edition 4. The data columns are:
    GNP, Invest, Price, Interest and computed realInvest, realGNP, 
    and realInterest. The Dates are ISO-8601 formatted as \%Y-\%m-%d.
       
    \code{pr45}
    \cr
    Sales and durable goods data from example 4.5 in the book of
    R.S. Pyndick and D.L. Rubinfeld. The data are seasonally adjusted
    variables from the Citibase data base: RTDR, IVRDR, FYCP, LEH,
    and PUCD.
  
    \code{spc1970} and 
    \code{spcindis}
    \cr
    These data are indicators for daily SP500 Stock Data together
    with an set of trading indicators for technical analysis.
    \cr
    
}


\format{
    
    All files are in CSV Excel spreadsheet format. The
    delimiter is a semicolon. 
    
}



\examples{
## SOURCE("fMultivar.048A-MultivarData")

## -
} 
 
 
\keyword{datasets}

