\name{MvtnormTools}


\alias{MvtnormTools}


\title{Mvtnorm Tools}


\description{

    Builtin functions from R's contributed Mvtnorm package.
    \cr
    
}


\author{

    Friedrich Leisch for the package Mvtnorm, \cr
    Diethelm Wuertz for the Rmetrics \R-port.

}


\examples{
## SOURCE("fMultivar.9B-MvtnormBuiltin")

## -
}


\keyword{programming}

